inspektor: Move modules to utils
As they are utility libraries, let's move stacktrace and
others to the utils namespace.
