+++ inspektor/utils/stacktrace.py
